finish most part of graphql and db
