+++ controllers/order.js
@@ -1,103 +1,7 @@
-import {pubsub} from "../resolvers/context.js"
+import { pubsub } from "../resolvers/context.js"
+import { orderList } from "../db/db.js"
+import {dbQuery} from "../db/connection.js"
 
-let orderList = [
-    {
-        "id": "order001",
-        "tableNo": "15A",
-        "totalPrice": 165,
-        "time": "2022-04-11T12:20:05",
-        "items": [
-            {
-                "id": "item001",
-                "name": "蛋包飯",
-                "price": 80,
-                "quantity": 1,
-                "note": "飯不要太多",
-                "status": "RAW"
-            },
-            {
-                "id": "item010",
-                "name": "珍珠鮮奶茶",
-                "price": 60,
-                "quantity": 1,
-                "note": null,
-                "status": "RAW"
-            },
-            {
-                "id": "item007",
-                "name": "泡芙",
-                "price": 25,
-                "quantity": 1,
-                "note": null,
-                "status": "RAW"
-            }
-        ]
-    },
-    {
-        "id": "order002",
-        "tableNo": "12D",
-        "totalPrice": 765,
-        "time": "2022-04-11T13:15:12",
-        "items": [
-            {
-                "id": "item002",
-                "name": "義大利麵",
-                "price": 100,
-                "quantity": 4,
-                "note": "兩個青醬兩個紅醬",
-                "status": "RAW"
-            },
-            {
-                "id": "item004",
-                "name": "水果茶",
-                "price": 90,
-                "quantity": 2,
-                "note": null,
-                "status": "RAW"
-            },
-            {
-                "id": "item009",
-                "name": "薯條",
-                "price": 40,
-                "quantity": 2,
-                "note": "要番茄醬",
-                "status": "RAW"
-            },
-            {
-                "id": "item008",
-                "name": "洋蔥圈",
-                "price": 45,
-                "quantity": 1,
-                "note": null,
-                "status": "RAW"
-            },
-            {
-                "id": "item010",
-                "name": "珍珠鮮奶茶",
-                "price": 60,
-                "quantity": 1,
-                "note": null,
-                "status": "RAW"
-            }
-        ]
-    },
-    {
-        "id": "order003",
-        "tableNo": 3,
-        "totalPrice": 120,
-        "time": "2022-04-11T14:01:18",
-        "items": [
-            {
-                "id": "item006",
-                "name": "披薩",
-                "price": 120,
-                "quantity": 1,
-                "note": null,
-                "status": "RAW"
-            }
-        ]
-    }
-]
 
 const postOrder = (req, res)=>{
     console.log("postOrder got ", req.body)
@@ -107,8 +11,18 @@ const postOrder = (req, res)=>{
     })
     res.status(200).send("success")
 }
-const getAllOrder = (req, res)=>{
-    res.status(200).send(orderList)
+const getAllOrder = async (req, res)=>{
+    let queryString = `SELECT * FROM \`Order\``
+    let orderResult = await dbQuery(queryString)
+    let orderItemResult = await dbQuery('SELECT * FROM `Order_Item`')
+    let itemResult = await dbQuery('SELECT * FROM `Item`')
+    let itemTransResult = await dbQuery('SELECT * FROM `Item_Trans`')
+    orderItemResult.forEach(e => {
+        // TODO LOGIC PART
+    })
+    
+    res.status(200).send({orderResult, orderItemResult, itemResult, itemTransResult})
+
 }
 const addToOrderList = (element)=>{
     orderList = [...orderList, element]
