+++ resolvers/Mutation.js
@@ -1,5 +1,6 @@
 import {pubsub} from "./context.js"
 import { orderList, addToOrderList } from "../controllers/order.js"
+import {dbMutation} from "../db/connection.js"
 // ! passing context from Appolo server constructor is not working
 const Mutation = {
     createOrder(parent, {order}, {}, info){
@@ -11,11 +12,25 @@ const Mutation = {
         })
         return "success"
     },
-    broadcastRandomNumber: (parent, {order}, {db, pubSub}, info) => {
-        // publish a random number
-        pubsub.publish('randomNumber', 1.4)
-        return 1.2
-      },
+    async createItem(parent, {data}, {}, info){
+        console.log("recevied data:", data)
+        let itemId = "item"+Math.floor(Math.random()*1000)
+        console.log(typeof(itemId))
+        let result = await dbMutation(`INSERT INTO \`Item\` VALUES('${itemId}', '${data.img}', ${data.price} )`)
+        //! schema input type is not well defined, so we cannot insert full info into db
+        result = await dbMutation(`INSERT INTO \`Item_Trans\` VALUES('${Math.floor(Math.random()*1000)}', '${data.name}' ,'zh', '', '', '${itemId}')`)
+        console.log(result)
+        return data
+    },
+    updateItem(parent, {id, data}, {}, info){
+        console.log("id:", id)
+        console.log("recevied data:", data)
+        // todo
+    },
+    deleteItem(parent, {id, data}, {}, info){
+        console.log("id:", id)
+        // todo
+    }
 }
 
 export default Mutation
